Possibilité pour les professeurs de changer la deadline.

+++ resources/views/groups/list.blade.php
@@ -7,7 +7,9 @@
 	<div class="panel panel-default">
 		<div class="panel-heading">Liste des classes</div>
 		<div class="panel-body">
-			<a href="classes/create"><button class="btn btn-primary pull-right">Ajouter une classe</button></a>
+			@can('create', App\Models\Group::class)
+				<a href="classes/create"><button class="btn btn-primary pull-right">Ajouter une classe</button></a>
+			@endcan
 			<table class="table table-striped">
 				<thead>
 					<tr>
@@ -25,19 +27,26 @@
 							<td>{{$group->year}}</td>
 							<td>@include('groups.partials.lock')</td>
 							<td>{{$group->course->name}}</td>
-							<td><a href="classes/{{$group->id}}">
-								<button class="btn btn-primary actionButton"><i class="fa fa-eye"></i></button>
-							</a>
-							<a href="classes/{{$group->id}}/edit">
-								<button class="btn btn-warning actionButton"><i class="fa fa-edit"></i></button>
-							</a>
-							{{Form::open(['method' => 'delete',
-										'url' => 'classes/'.$group->id,
-										'class' => 'deleteBtn'])}}
-								<button type="submit" class="btn btn-danger actionButton">
-									<i class="fa fa-trash"></i>
-								</button>
-							{{Form::close()}}
+							<td>
+									<a href="classes/{{$group->id}}">
+										@can('view', $group)
+										<button class="btn btn-primary actionButton"><i class="fa fa-eye"></i></button>
+									</a>
+								@endcan
+								@can('edit', $group)
+									<a href="classes/{{$group->id}}/edit">
+										<button class="btn btn-warning actionButton"><i class="fa fa-edit"></i></button>
+									</a>
+								@endcan
+								@can('delete', $group)
+									{{Form::open(['method' => 'delete',
+												'url' => 'classes/'.$group->id,
+												'class' => 'deleteBtn'])}}
+										<button type="submit" class="btn btn-danger actionButton">
+											<i class="fa fa-trash"></i>
+										</button>
+									{{Form::close()}}
+								@endcan
 							</td>
 						</tr>
 					@endforeach
